Add support for auto-detecting FBE/FDE/Decrypted
In order to use this feature, make sure you remove any of the
following flags from your fstab:

  encryptable=, fileencryption=, length=

And add the flag autodetectcrypt

How it works:

It checks whether the FDE footer exists at -16384 (16384
from the end of the partition). If that is the case,
it does the same that would happen if encryptable=footer,length=-16384
was specified in fstab. Otherwise it will check if you have FBE, that means
it will read-only mount data at a temporary place and check whether the
FBE key exists. If it does, it will do the same that
fileencryption=ice does, otherwise it will just behave like encryption was
disabled.

Change-Id: I50253c52ec0105809c696c8fd94df1b5ca6fe7cf
Signed-off-by: Simao Gomes Viana <xdevs23@outlook.com>